Remove unused (and duplicate) method
