+++ api/tests/unit/domain/services/certification/certification-result-service_test.js
@@ -4,6 +4,7 @@ const certificationResultService = require('../../../../../lib/domain/services/c
 
 const Assessment = require('../../../../../lib/domain/models/Assessment');
 const CertificationCourse = require('../../../../../lib/domain/models/CertificationCourse');
+const Answer = require('../../../../../lib/domain/models/Answer');
 
 const answersRepository = require('../../../../../lib/infrastructure/repositories/answer-repository');
 const certificationChallengesRepository = require('../../../../../lib/infrastructure/repositories/certification-challenge-repository');
@@ -1223,6 +1224,77 @@ describe('Unit | Service | Certification Result Service', function() {
       });
     });
 
+  });
+
+  describe('#computeAnswersSuccessRate', () => {
+
+    context('when no answers is given', () => {
+
+      it('should have a trust level has unknown', () => {
+        // given
+        const answers = [];
+
+        // when
+        const successRate = certificationResultService._computeAnswersSuccessRate(answers);
+
+        // then
+        expect(successRate).to.equal(0);
+      });
+    });
+
+    context('when all answers are OK', () => {
+
+      it('should have a success rate of 100%', () => {
+        // given
+        const answers = [new Answer({ result: 'ok' }), new Answer({ result: 'ok' })];
+
+        // when
+        const successRate = certificationResultService._computeAnswersSuccessRate(answers);
+
+        // then
+        expect(successRate).to.equal(100);
+      });
+    });
+
+    context('when all answers are KO', () => {
+
+      it('should have a success rate of 0%', () => {
+        // given
+        const answers = [new Answer({ result: 'ko' }), new Answer({ result: 'ko' })];
+
+        // when
+        const successRate = certificationResultService._computeAnswersSuccessRate(answers);
+
+        // then
+        expect(successRate).to.equal(0);
+      });
+    });
+
+    context('when the answers are a mixed of valid and wrong answers', () => {
+
+      it('should have a success rate of 50% with 1W and 1R', () => {
+        // given
+        const answers = [new Answer({ result: 'ok' }), new Answer({ result: 'ko' })];
+
+        // when
+        const successRate = certificationResultService._computeAnswersSuccessRate(answers);
+
+        // then
+        expect(successRate).to.equal(50);
+      });
+
+      it('should have a success rate of 50% with 2W and 1R', () => {
+        // given
+        const answers = [new Answer({ result: 'ok' }), new Answer({ result: 'aband' }), new Answer({ result: 'ko' })];
+
+        // when
+        const successRate = certificationResultService._computeAnswersSuccessRate(answers);
+
+        // then
+        expect(successRate).to.be.within(33.333333, 33.333334);
+      });
+
+    });
   });
 
 });
